+++ dev-site-config/site.config.js
@@ -30,8 +30,6 @@ const siteConfig = {
       },
     ],
   },
-
-  includeTestEvidence: false,
 };
 
 module.exports = siteConfig;