+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Breaking
+  * Remove the ability to include test evidence, this is covered now by a separate site report.
+
 ## 6.30.0 - (September 11, 2020)
 
 * Changed