+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Breaking
   * Remove the ability to include test evidence, this is covered now by a separate site report.
+  * Use MDX instead of Marked to load markdown files.
 
 ## 6.30.0 - (September 11, 2020)